feat(): rename file name
Refs: NEXT_BUILDER-1113
